frontend/backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
supabase==2.3.4
PyPDF2==3.0.1
numpy==1.24.3
sentence-transformers==2.5.1
huggingface-hub==0.20.3
rq==1.15.1
python-multipart==0.0.9 